Move Simple table to common
